Refactor to use Duck pattern
