+++ src/components/Muffins/Muffins.js
@@ -2,11 +2,11 @@ import React, { useEffect } from 'react';
 
 import { useSelector, useDispatch  } from 'react-redux';
 
-import { selectedMuffinsArray, selectMuffinsLoadError, selectMuffinsLoading } from '../../redux/selectors';
-import { likeMuffin, loadMuffins  } from '../../redux/actions';
+import { selectMuffinsArray, selectMuffinsLoadError, selectMuffinsLoading } from '../../redux/muffins';
+import { likeMuffin, loadMuffins  } from '../../redux/muffins';
 
 const Muffins = () => {
-  const muffins = useSelector(selectedMuffinsArray);
+  const muffins = useSelector(selectMuffinsArray);
   const muffinsLoading = useSelector(selectMuffinsLoading);
   const loadError = useSelector(selectMuffinsLoadError);
