--- src/redux/selectors.js
@@ -1,3 +1,0 @@
-export const selectedMuffinsArray = (state) => state.muffins;
-export const selectMuffinsLoading = (state) => state.selectMuffinsLoading;
-export const selectMuffinsLoadError = (state) => state.error;
